added comments for public and private classes
